tweaked file structure
I moved the file structure around a bit to reflect what we had going on
in class. The only assets i moved were the two images inside the
innermost "assets" folder. I've added two folders, images, and sounds,
at the same level as the "flash" folder. I also added these two new
folders to the class path for the project. from now on, we should put
all assets in one of these two folders (assets that will be used in the
game). This should help us keep track of what images we are using, and
which ones we aren't using.